Move npc interfaces to api module
