+++ code/api/src/main/java/org/betonquest/betonquest/api/quest/npc/NpcRegistry.java
