+++ code/compatibility/src/main/java/org/betonquest/betonquest/compatibility/mythicmobs/npc/MythicMobsInteractCatcher.java
@@ -8,8 +8,8 @@ import io.lumine.mythic.core.mobs.MobExecutor;
 import io.papermc.paper.event.entity.EntityMoveEvent;
 import org.betonquest.betonquest.api.bukkit.event.npc.NpcVisibilityUpdateEvent;
 import org.betonquest.betonquest.api.profile.ProfileProvider;
+import org.betonquest.betonquest.api.quest.npc.NpcInteractCatcher;
 import org.betonquest.betonquest.api.quest.npc.NpcRegistry;
-import org.betonquest.betonquest.api.quest.npc.feature.NpcInteractCatcher;
 import org.betonquest.betonquest.compatibility.mythicmobs.MythicHider;
 import org.betonquest.betonquest.quest.objective.interact.Interaction;
 import org.bukkit.entity.Player;
