+++ code/compatibility/src/main/java/org/betonquest/betonquest/compatibility/npc/citizens/CitizensInventoryConvIO.java
@@ -8,7 +8,7 @@ import org.betonquest.betonquest.api.config.quest.QuestPackageManager;
 import org.betonquest.betonquest.api.logger.BetonQuestLogger;
 import org.betonquest.betonquest.api.profile.OnlineProfile;
 import org.betonquest.betonquest.api.quest.Placeholders;
-import org.betonquest.betonquest.api.quest.npc.feature.NpcConversation;
+import org.betonquest.betonquest.api.quest.npc.NpcConversation;
 import org.betonquest.betonquest.conversation.Conversation;
 import org.betonquest.betonquest.conversation.ConversationColors;
 import org.betonquest.betonquest.conversation.InventoryConvIO;
