+++ code/compatibility/src/main/java/org/betonquest/betonquest/compatibility/npc/fancynpcs/FancyCatcher.java
@@ -6,8 +6,8 @@ import de.oliver.fancynpcs.api.events.NpcInteractEvent;
 import de.oliver.fancynpcs.api.events.NpcModifyEvent;
 import org.betonquest.betonquest.api.bukkit.event.npc.NpcVisibilityUpdateEvent;
 import org.betonquest.betonquest.api.profile.ProfileProvider;
+import org.betonquest.betonquest.api.quest.npc.NpcInteractCatcher;
 import org.betonquest.betonquest.api.quest.npc.NpcRegistry;
-import org.betonquest.betonquest.api.quest.npc.feature.NpcInteractCatcher;
 import org.betonquest.betonquest.quest.objective.interact.Interaction;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
