+++ code/compatibility/src/main/java/org/betonquest/betonquest/compatibility/npc/fancynpcs/FancyHider.java
@@ -1,7 +1,7 @@
 package org.betonquest.betonquest.compatibility.npc.fancynpcs;
 
 import de.oliver.fancynpcs.api.events.NpcSpawnEvent;
-import org.betonquest.betonquest.api.quest.npc.feature.NpcHider;
+import org.betonquest.betonquest.api.quest.npc.NpcHider;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
