+++ code/core/src/main/java/org/betonquest/betonquest/api/quest/npc/NpcConversation.java
@@ -1,10 +1,9 @@
-package org.betonquest.betonquest.api.quest.npc.feature;
+package org.betonquest.betonquest.api.quest.npc;
 
 import org.betonquest.betonquest.api.QuestException;
 import org.betonquest.betonquest.api.identifier.ConversationIdentifier;
 import org.betonquest.betonquest.api.logger.BetonQuestLogger;
 import org.betonquest.betonquest.api.profile.OnlineProfile;
-import org.betonquest.betonquest.api.quest.npc.Npc;
 import org.betonquest.betonquest.config.PluginMessage;
 import org.betonquest.betonquest.conversation.Conversation;
 import org.bukkit.Location;
