+++ code/core/src/main/java/org/betonquest/betonquest/api/quest/npc/NpcInteractCatcher.java
@@ -1,11 +1,9 @@
-package org.betonquest.betonquest.api.quest.npc.feature;
+package org.betonquest.betonquest.api.quest.npc;
 
 import org.betonquest.betonquest.api.bukkit.event.npc.NpcInteractEvent;
 import org.betonquest.betonquest.api.identifier.NpcIdentifier;
 import org.betonquest.betonquest.api.profile.OnlineProfile;
 import org.betonquest.betonquest.api.profile.ProfileProvider;
-import org.betonquest.betonquest.api.quest.npc.Npc;
-import org.betonquest.betonquest.api.quest.npc.NpcRegistry;
 import org.betonquest.betonquest.quest.objective.interact.Interaction;
 import org.bukkit.entity.Player;
 import org.bukkit.event.Listener;
