+++ code/core/src/main/java/org/betonquest/betonquest/quest/action/npc/UpdateVisibilityNowAction.java
@@ -2,7 +2,7 @@ package org.betonquest.betonquest.quest.action.npc;
 
 import org.betonquest.betonquest.api.profile.OnlineProfile;
 import org.betonquest.betonquest.api.quest.action.OnlineAction;
-import org.betonquest.betonquest.api.quest.npc.feature.NpcHider;
+import org.betonquest.betonquest.api.quest.npc.NpcHider;
 
 /**
  * Action to update the visibility of all Npcs to one player now.
