+++ code/core/src/main/java/org/betonquest/betonquest/quest/action/npc/UpdateVisibilityNowActionFactory.java
@@ -5,7 +5,7 @@ import org.betonquest.betonquest.api.logger.BetonQuestLoggerFactory;
 import org.betonquest.betonquest.api.quest.action.OnlineActionAdapter;
 import org.betonquest.betonquest.api.quest.action.PlayerAction;
 import org.betonquest.betonquest.api.quest.action.PlayerActionFactory;
-import org.betonquest.betonquest.api.quest.npc.feature.NpcHider;
+import org.betonquest.betonquest.api.quest.npc.NpcHider;
 
 /**
  * Factory to create {@link UpdateVisibilityNowAction}s from {@link Instruction}s.
